pyproject.toml
[tool.poetry]
name = "url-shortener"
version = "0.1.0"
description = "Simple app to shorten provided urls (and to get long urls back based on the short code)"
authors = ["Łukasz <lukasz@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.6"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
